+++ include/word.h
@@ -16,7 +16,7 @@ public:
 	bool parse_runes(const std::string& str, std::vector<Rune>& runes);
 	bool decode_image(const cv::Mat& word_image);
 	size_t size() const { return m_runes.size(); }
-	bool generate_image(cv::Size2i rune_size, int tickness, cv::Mat& output_image) const;
+	bool generate_image(cv::Size2i rune_size, double tickness, cv::Mat& output_image) const;
 	std::vector<Rune> get_runes() const { return m_runes; }
 	bool is_valid() const { return !m_runes.empty(); }
 private:
